+++ packages/extensions/core/lib/configuration/configuration.ts
@@ -19,12 +19,11 @@ import {
   TryDecodeEnhancedPositionFromName,
 } from "@azure-tools/datastore";
 import { Extension, ExtensionManager, LocalExtension } from "@azure-tools/extension";
-import { clone, keys, Dictionary, values } from "@azure-tools/linq";
+import { clone, Dictionary, values } from "@azure-tools/linq";
 import {
   CreateFileUri,
   CreateFolderUri,
   EnsureIsFolderUri,
-  ExistsUri,
   ResolveUri,
   simplifyUri,
   IsUri,
@@ -35,18 +34,20 @@ import {
 import { From } from "linq-es2015";
 import { basename, dirname, join } from "path";
 import { CancellationToken, CancellationTokenSource } from "vscode-jsonrpc";
-import { Artifact } from "./artifact";
-import * as Constants from "./constants";
-import { EventEmitter, IEvent } from "./events";
-import { OperationAbortedException } from "./exception";
-import { Channel, Message, Range, SourceLocation } from "./message";
-import { evaluateGuard, parseCodeBlocks } from "./parsing/literate-yaml";
-import { AutoRestExtension } from "./pipeline/plugin-endpoint";
-import { Suppressor } from "./pipeline/suppression";
-import { MergeOverwriteOrAppend, resolveRValue } from "./source-map/merging";
-import { Initializer, DeepPartial } from "@azure-tools/codegen";
+import { Artifact } from "../artifact";
+import * as Constants from "../constants";
+import { EventEmitter, IEvent } from "../events";
+import { OperationAbortedException } from "../exception";
+import { Channel, Message, Range, SourceLocation } from "../message";
+import {  parseCodeBlocks } from "../parsing/literate-yaml";
+import { AutoRestExtension } from "../pipeline/plugin-endpoint";
+import { Suppressor } from "../pipeline/suppression";
+import {  resolveRValue } from "../source-map/merging";
+import { Initializer } from "@azure-tools/codegen";
 import { cwd } from "process";
-import { AppRoot } from "./constants";
+import { AppRoot } from "../constants";
+import { Directive } from "./directive";
+import { AutoRestConfigurationImpl, mergeConfiguration, mergeConfigurations } from "./auto-rest-configuration-impl";
 
 const safeEval = createSandbox();
 
@@ -54,97 +55,6 @@ const safeEval = createSandbox();
 const untildify: (path: string) => string = require("untildify");
 
 const RESOLVE_MACROS_AT_RUNTIME = true;
-
-export interface AutoRestConfigurationImpl {
-  "__info"?: string | null;
-  "__parents"?: any | undefined;
-  "allow-no-input"?: boolean;
-  "input-file"?: Array<string> | string;
-  "exclude-file"?: Array<string> | string;
-  "base-folder"?: string;
-  "directive"?: Array<Directive> | Directive;
-  "declare-directive"?: { [name: string]: string };
-  "output-artifact"?: Array<string> | string;
-  "message-format"?: "json" | "yaml" | "regular";
-  "use-extension"?: { [extensionName: string]: string };
-  "require"?: Array<string> | string;
-  "try-require"?: Array<string> | string;
-  "help"?: any;
-  "vscode"?: any; // activates VS Code specific behavior and does *NOT* influence the core's behavior (only consumed by VS Code extension)
-
-  "override-info"?: any; // make sure source maps are pulling it! (see "composite swagger" method)
-  "title"?: any;
-  "description"?: any;
-
-  "debug"?: boolean;
-  "verbose"?: boolean;
-  "time"?: boolean;
-
-  // plugin specific
-  "output-file"?: string;
-  "output-folder"?: string;
-
-  // from here on: CONVENTION, not cared about by the core
-  "client-side-validation"?: boolean; // C#
-  "fluent"?: boolean;
-  "azure-arm"?: boolean;
-  "namespace"?: string;
-  "license-header"?: string;
-  "add-credentials"?: boolean;
-  "package-name"?: string; // Ruby, Python, ...
-  "package-version"?: string;
-  "sync-methods"?: "all" | "essential" | "none";
-  "payload-flattening-threshold"?: number;
-  "openapi-type"?: string; // the specification type (ARM/Data-Plane/Default)
-  "tag"?: string;
-  "simple-tree-shake"?: boolean;
-
-  // multi-api specific
-  "profiles"?: any;
-  "profile"?: Array<string> | string;
-  "api-version"?: Array<string>;
-
-  "pipeline-model"?: string;
-  "load-priority"?: number;
-
-  "resolved-directive"?: any;
-  "debugger"?: any;
-
-  "github-auth-token"?: string;
-}
-
-export function MergeConfigurations(...configs: Array<AutoRestConfigurationImpl>): AutoRestConfigurationImpl {
-  let result: AutoRestConfigurationImpl = {};
-  configs = configs
-    .map((each, i, a) => ({ ...each, "load-priority": each["load-priority"] || -i }))
-    .sort((a, b) => b["load-priority"] - a["load-priority"]);
-  // if they say --profile: or --api-version: (or in config) then we force it to set the tag=all-api-versions
-  // Some of the rest specs had a default tag set (really shouldn't have done that), which ... was problematic,
-  // so this enables us to override that in the case they are asking for filtering to a profile or a api-verison
-
-  const forceAllVersionsMode = !!configs.find((each) => each["api-version"]?.length || each.profile?.length || 0 > 0);
-  for (const config of configs) {
-    result = MergeConfiguration(result, config, forceAllVersionsMode);
-  }
-  result["load-priority"] = undefined;
-  return result;
-}
-
-// TODO: operate on DataHandleRead and create source map!
-function MergeConfiguration(
-  higherPriority: AutoRestConfigurationImpl,
-  lowerPriority: AutoRestConfigurationImpl,
-  forceAllVersionsMode = false,
-): AutoRestConfigurationImpl {
-  // check guard
-  if (lowerPriority.__info && !evaluateGuard(lowerPriority.__info, higherPriority, forceAllVersionsMode)) {
-    // guard false? => skip
-    return higherPriority;
-  }
-
-  // merge
-  return MergeOverwriteOrAppend(higherPriority, lowerPriority);
-}
 
 function isIterable(target: any): target is Iterable<any> {
   return !!target && typeof target[Symbol.iterator] === "function";
@@ -197,18 +107,6 @@ function arrayOf<T>(value: any): Array<T> {
       break;
   }
   return [<T>value];
-}
-
-export interface Directive extends Dictionary<any> {
-  from?: Array<string> | string;
-  where?: Array<string> | string;
-  reason?: string;
-
-  // one of:
-  suppress?: Array<string> | string;
-  set?: Array<string> | string;
-  transform?: Array<string> | string;
-  test?: Array<string> | string;
 }
 
 export class ResolvedDirective extends Initializer implements Dictionary<any> {
@@ -346,11 +244,11 @@ export class ConfigurationView {
       "pass-thru": [],
     };
 
-    this.rawConfig = MergeConfigurations(this.rawConfig, ...configs);
+    this.rawConfig = mergeConfigurations(this.rawConfig, ...configs);
 
     // default values that are the least priority.
     // TODO: why is this here and not in default-configuration?
-    this.rawConfig = MergeConfiguration(this.rawConfig, <any>{
+    this.rawConfig = mergeConfiguration(this.rawConfig, <any>{
       "base-folder": ".",
       "output-folder": "generated",
       "debug": false,
