+++ packages/extensions/core/lib/configuration/index.ts
@@ -1,0 +1,2 @@
+export * from "./configuration";
+export * from "./auto-rest-configuration-impl";
